Fix feedback from PHPStan and PHP CS Fixer

+++ src/Adapter/Store/CommandHandler/EditStoreHandler.php
@@ -44,10 +44,10 @@ final class EditStoreHandler implements EditStoreHandlerInterface
             $store->postcode = $command->getPostcode();
         }
         if (null !== $command->getLatitude()) {
-            $store->latitude = number_format($command->getLatitude(), 8, '.', '');
+            $store->latitude = (float) number_format($command->getLatitude(), 8, '.', '');
         }
         if (null !== $command->getLongitude()) {
-            $store->longitude = number_format($command->getLongitude(), 8, '.', '');
+            $store->longitude = (float) number_format($command->getLongitude(), 8, '.', '');
         }
         if (null !== $command->getPhone()) {
             $store->phone = $command->getPhone();
